added login with github

+++ app/Http/Controllers/Profile/AvatarController.php
@@ -5,13 +5,16 @@ namespace App\Http\Controllers\Profile;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UpdateAvatarRequest;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
 
 class AvatarController extends Controller
 {
     public function update(UpdateAvatarRequest $request)
     {
         if($currentAvatar = request()->user()->avatar){
-            Storage::disk('public')->delete($currentAvatar);
+            if(!Str::contains($currentAvatar, 'avatars/default-user.jpg')){
+                Storage::disk('public')->delete($currentAvatar);
+            }
         }
 
         $path = Storage::disk('public')->put('avatars', $request->file('avatar'));
